+++ main/Display/dispgrob.cc
@@ -59,6 +59,7 @@ static const char* Copyright = "(C) Copyright Michigan State University 1994, Al
 #include <iostream>
 
 #include <math.h>
+#include <stdint.h>
 
 #ifdef HAVE_STD_NAMESPACE
 using namespace std;
@@ -402,7 +403,7 @@ void grobj_sum1d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  long index, row, col;
+  uintptr_t index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -475,7 +476,7 @@ void grobj_cut1d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  long index, row, col;
+  uintptr_t index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -548,7 +549,7 @@ void grobj_mark1d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  long index, row, col;
+  uintptr_t index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -623,7 +624,7 @@ void grobj_contour::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  long index, row, col;
+  uintptr_t index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -702,7 +703,7 @@ void grobj_band::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  long index, row, col;
+  uintptr_t index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -778,7 +779,7 @@ void grobj_sum2d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  long index, row, col;
+  uintptr_t index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -857,7 +858,7 @@ void grobj_mark2d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  long index, row, col;
+  uintptr_t index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -1019,7 +1020,7 @@ void grobj_ptlist1::draw(XMWidget *p, win_attributed *at, Boolean final)
   **  so that if the window is contaiminated while accepting points, the 
   **  partial point set will be visible.
   */
-  long index, row, col;
+  uintptr_t index, row, col;
   p->GetAttribute(XmNuserData, &index); // This is the pane number.
   row = index % WINDOW_MAXAXIS;
   col = index / WINDOW_MAXAXIS;
@@ -1153,7 +1154,7 @@ void grobj_ptlist2::draw(XMWidget *p, win_attributed *at, Boolean final)
   **  so that if the window is contaiminated while accepting points, the 
   **  partial point set will be visible.
   */
-  long index, row, col;
+  uintptr_t index, row, col;
   p->GetAttribute(XmNuserData, &index); // This is the pane number.
   row = index % WINDOW_MAXAXIS;
   col = index / WINDOW_MAXAXIS;
@@ -1257,7 +1258,7 @@ void grobj_Peak1d::draw(XMWidget* pWindow,
   // we rely on the fact that the pane widget has the pane number in it's
   // user data field:
 
-  long index, row, col;
+  uintptr_t index, row, col;
   pWindow->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -1723,7 +1724,7 @@ grobj_Fitline::draw(XMWidget* pWindow, win_attributed* pAttributes,
   // we rely on the fact that the pane widget has the pane number in it's
   // user data field:
 
-  long index, row, col;
+  uintptr_t index, row, col;
   pWindow->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
